Move `initInstance` call into `loadModule[A]sync`
Currently, every call to `loadModuleSync` or `loadModuleAsync` in the node
entry points has to be followed by a call to `initInstance`. This change moves
that call into `loadModule[A]sync` instead.

This (a) brings the node code more into line with the browser code; (b) lays
groundwork for the future where I want to move bits of `initInstance` into `loadModule[A]sync`.